pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hermes-jepa-edge"
version = "0.1.0"
description = "JEPA world model + Hermes LLM edge robotic agent"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.3.0",
    "torchvision>=0.18.0",
    "transformers>=4.45.0",
    "einops>=0.8.0",
    "huggingface-hub>=0.24.0",
    "datasets>=2.20.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "numpy>=1.26.0",
    "tqdm>=4.66.0",
    "dataclasses-json>=0.6.0",
]

[project.optional-dependencies]
train = [
    "flash-attn>=2.6.0",
    "bitsandbytes>=0.43.0",   # QLoRA (NF4 quantization)
    "peft>=0.12.0",            # LoRA adapter injection + merge
    "accelerate>=0.33.0",
    "trl>=0.10.0",             # SFTTrainer for distillation
    "wandb>=0.17.0",
]
edge = [
    "onnx>=1.16.0",
    "onnxruntime-gpu>=1.19.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
